Changed private vars to 1 byte

+++ Libraries/MCP23S08/MCP23S08.h
@@ -84,16 +84,16 @@ class MCP23S08 {
     MCP23S08(uint8_t, uint8_t);              // Constructor to instantiate a discrete IC as an object, address 0-7, chipSelect any valid pin
     MCP23S08(uint8_t, uint8_t, unsigned int);
     void begin();                            // Start the SPI Bus and write default values to all registers
-    void wordWrite(uint8_t, unsigned int);   // Typically only used internally, but allows the user to write any register pair if needed, so it's public
+    void wordWrite(uint8_t, uint8_t);        // Typically only used internally, but allows the user to write any register pair if needed, so it's public
     void byteWrite(uint8_t, uint8_t);        // Typically only used internally, but allows the user to write any register if needed, so it's public
     void pinMode(uint8_t, uint8_t);          // Sets the mode (input or output) of a single I/O pin
-    void pinMode(unsigned int);              // Sets the mode (input or output) of all I/O pins at once
+    void pinMode(uint8_t);                   // Sets the mode (input or output) of all I/O pins at once
     void pullupMode(uint8_t, uint8_t);       // Selects internal 100k input pull-up of a single I/O pin
-    void pullupMode(unsigned int);           // Selects internal 100k input pull-up of all I/O pins at once
+    void pullupMode(uint8_t);                // Selects internal 100k input pull-up of all I/O pins at once
     void inputInvert(uint8_t, uint8_t);      // Selects input state inversion of a single I/O pin (writing 1 turns on inversion)
-    void inputInvert(unsigned int);          // Selects input state inversion of all I/O pins at once (writing a 1 turns on inversion)
+    void inputInvert(uint8_t);               // Selects input state inversion of all I/O pins at once (writing a 1 turns on inversion)
     void digitalWrite(uint8_t, uint8_t);     // Sets an individual output pin HIGH or LOW
-    void digitalWrite(unsigned int);         // Sets all output pins at once. If some pins are configured as input, those bits will be ignored on write
+    void digitalWrite(uint8_t);              // Sets all output pins at once. If some pins are configured as input, those bits will be ignored on write
     uint8_t digitalRead(uint8_t);            // Reads an individual input pin
     uint8_t byteRead(uint8_t);               // Reads an individual register and returns the byte. Argument is the register address
     unsigned int digitalRead(void);          // Reads all input  pins at once. Be sure it ignore the value of pins configured as output!
@@ -101,11 +101,11 @@ class MCP23S08 {
     void init(uint8_t, uint8_t, unsigned int);
     uint8_t _address;                        // Address of the MCP23S08 in use
 	uint8_t _cs;                             // chip select pin
-    unsigned int _modeCache;                 // Caches the mode (input/output) configuration of I/O pins
-    unsigned int _pullupCache;               // Caches the internal pull-up configuration of input pins (values persist across mode changes)
-    unsigned int _invertCache;               // Caches the input pin inversion selection (values persist across mode changes)
-    unsigned int _outputCache;               // Caches the output pin state of pins
-    unsigned int _SPIspeed;
+    uint8_t int _modeCache;                 // Caches the mode (input/output) configuration of I/O pins
+    uint8_t int _pullupCache;               // Caches the internal pull-up configuration of input pins (values persist across mode changes)
+    uint8_t int _invertCache;               // Caches the input pin inversion selection (values persist across mode changes)
+    uint8_t int _outputCache;               // Caches the output pin state of pins
+    uint8_t int _SPIspeed;
  };
 
 #endif //MCP23S08
